Fix full flag mode with Mac.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@8595 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/java/jmsg.c
@@ -251,7 +251,7 @@ Java_simgrid_msg_MsgNative_processFromPID(JNIEnv * env, jclass cls,
   m_process_t process = MSG_process_from_PID(PID);
 
   if (!process) {
-    jxbt_throw_process_not_found(env, bprintf("PID = %d", PID));
+    jxbt_throw_process_not_found(env, bprintf("PID = %d",(int) PID));
     return NULL;
   }
